+++ tests/ApplicationTest.php
@@ -47,6 +47,8 @@ class ApplicationTest extends TestCase
 
         $this->assertInstanceOf(Application::class, $application);
 
+        $this->assertSame(__DIR__.'/stubs', $application->getBasePath());
+
         $application->setPublicPath(__DIR__);
         $this->assertSame(__DIR__, $application->getPublicPath());